Update project to React (#10)
* (deps) install react and typescript dependencies

* add tsconfig

* update indent_size

* add react to vite config

* add react to base blade file

* update VITE env variables

* (deps) remove lodash

* (deps) bump axios to v1

* update app and bootstrap files

* add type files

* change welcome page to react

* update login page to react

* update register page to react

* update forgot password page to react

* update confirm password page to react

* update reset password page to react

* update verify email page to react

* (deps) install headlessui and heroicons for react

* update user dashboard to react

* update profile pages to react

* update editor dashboard to react

* update admin dashboard to react

* update content index page to react

* update collections index page to react

* update collection show page to react

* update project show page to react

* update assignments index page to react

* partial update of content related files to react

* update drafts show page to react

* beginning to update draft creation to react

* fix the assignment of sorted projects, goals, and choices

* remove vue drafts create page

* fix display of up and down arrows for reordering

* fix dropdown menu for secondary button with dropdown component

* add prose component

* update collections edit page to react

* update notification component to react

* update project edit page to react

* fix secondary small button to use onClick and fix table head columns

* update goal edit page to react

* update draft edit page to react

* update primary button to be a link if the href prop is present

* update danger modal

* add typing

* only allow admins to archive collections

* update notification

* update assignment show page to react

* update error layout and expired error page to react

* update editors index page to react

* update create editor invitation page to react

* update editor show page to react

* update editor edit page to react

* add danger modal to editor show page

* update invitation accept page to react

* update assignment edit page to react

* update assignment create page to react

* remove index and create views and routes for collections, projects, and goals, and choices

* remove outstanding vue layout files

* update active state of admin navigation

* update notification dropdown to react

* update typing for admin dashboard components

* update content card for content index page

* add expandable choice items to goal show page

* remove pages for choices

* get tiptap editor working

* update text block component to reaect

* aupdate header component to react

* update exercises component to react

* update quiklinks component to react

* add front end add new resource process

* add ability to create new resources

* save highlights to a video resource

* persist updates to a choice

* resources can be edited

* display danger modal to confirm deletion of a choice content property item

* add buttons for adding new content to a choice

* allow choice content items to be removed from the content

* store choice drafts

* update grid sizing for draft form

* move add, update, and delete methods for choice content into choice content form

* update grid sizing for description lists

* send updates to choice edit component from choice content form

* update focus styling for text editor

* display resources correctly on show

* update goal preview to react

* remove remaining vue files

* remove app.js vue setup

* fix build errors

* (pint) update php code styling

+++ app/Http/Controllers/CollectionController.php
@@ -4,6 +4,7 @@ namespace App\Http\Controllers;
 
 use App\Http\Requests\CollectionRequest;
 use App\Models\Collection;
+use Illuminate\Http\RedirectResponse;
 use Illuminate\Http\Request;
 
 class CollectionController extends Controller
@@ -15,9 +16,7 @@ class CollectionController extends Controller
      */
     public function index()
     {
-        return inertia('Editor/Content/Collections/Index', [
-            'collections' => Collection::all(),
-        ]);
+        //
     }
 
     /**
@@ -27,7 +26,7 @@ class CollectionController extends Controller
      */
     public function create()
     {
-        return inertia('Editor/Content/Collections/Create');
+        //
     }
 
     /**
@@ -99,11 +98,22 @@ class CollectionController extends Controller
     /**
      * Remove the specified resource from storage.
      *
-     * @param  \App\Models\Collection  $collection
-     * @return \Illuminate\Http\Response
+     * @param  Request  $request
+     * @param  Collection  $collection
+     * @return RedirectResponse
      */
-    public function destroy(Collection $collection)
+    public function destroy(Request $request, Collection $collection): RedirectResponse
     {
+        if ($request->user()->cannot('delete', $collection)) {
+            return redirect()->back()
+                ->with('notification', [
+                    'actions' => false,
+                    'message' => 'You do not have the required permissions to archive this collection.',
+                    'title' => 'Access error',
+                    'type' => 'error',
+                ]);
+        }
+
         $collection->delete();
 
         return redirect()->route('editor.collections.index');
